rktest.getTempDir {rkwardtests}	R Documentation

Get path to the temporary directory

Description

Get the path to the recent temporary directory, if one exists.

Usage

  rktest.getTempDir()

Details

This function will return either the local path to the temporary directory where all test results have been saved to, or FALSE if none exitsts.

Value

Either a character string, or FALSE.

Author(s)

Meik Michalke meik.michalke@uni-duesseldorf.de

See Also

rktest.makeplugintests

Examples

rktest.getTempDir()

[Package rkwardtests version 0.6.1 Index]