rktest.makeplugintests {rkwardtests}	R Documentation

Run RKWard plugin test suite

Description

Run a whole RKWard plugin test suite

Usage

  rktest.makeplugintests(testsuites="testsuite.R",
    testroot=getwd(), outfile="make_plugintests.txt",
    append=FALSE, test.id=NULL)

Arguments

testsuites
A character string or vector naming the test suites to be run.

testroot
A character string pointing to the root directory where the test suite resides (including its folder with test standards).

outfile
A character string giving a file name for the result log.

append
If TRUE, append output to an existing file.

test.id
Optional character string or vector naming one or more tests of a suite to be run (if NULL, all tests are run).

Details

The function rktest.makeplugintests will run a whole test suite that was prepared to check one or several RKWard plugins.

Value

Results are printed to stdout and saved to the defined output file.

Author(s)

Thomas Friedrichsmeier thomas.friedrichsmeier@ruhr-uni-bochum.de, Meik Michalke meik.michalke@uni-duesseldorf.de

See Also

RKTestSuite-class, RKTestResult-class

Examples

## Not run: 
rktest.makeplugintests(testsuites=c("rkward_application_tests.R",
   "import_export_plugins.R"), testroot=getwd())
rktest.makeplugintests(testsuites="distribution.R",
  testroot=getwd(), test.id=c("poisson_quantiles", "geom_quantiles"))

## End(Not run)

[Package rkwardtests version 0.6.1 Index]