rktest.replaceRunAgainLink {rkwardtests}	R Documentation

Replace "Run again" link in RKWard

Description

Replace "Run again" link in RKWard with code

Usage

  rktest.replaceRunAgainLink(restore=FALSE)

Arguments

restore
Logical: If TRUE, restore the original behaviour.

Details

You can use this to temporarily replace .rk.rerun.plugin.link (see example below). This way, after running a plugin, you are shown the call needed to run this plugin with those settings, instead of the link.

This code can be used in a plugin test suite.

Value

Replaces the "Run again" link in RKWard with the code that would have been called, or vice versa.

Author(s)

Thomas Friedrichsmeier thomas.friedrichsmeier@ruhr-uni-bochum.de, Meik Michalke meik.michalke@uni-duesseldorf.de

See Also

RKTestSuite-class, rktest.makeplugintests

Examples

rktest.replaceRunAgainLink()

[Package rkwardtests version 0.6.1 Index]