rktest.runRKTestSuite {rkwardtests}	R Documentation

Run RKWard plugin test suite

Description

Run a single RKWard plugin test suite

Usage

  rktest.runRKTestSuite(suite, testroot=getwd(),
    test.id=NULL)

Arguments

suite
Character string naming the test suite to run.

testroot
Defaults to the working directory.

test.id
An optional character string or vector naming one or more tests of a suite to be run (if NULL, all tests are run).

Details

This function can be called to run a single plugin test suite.

Value

An object of class RKTestResult-class.

Author(s)

Thomas Friedrichsmeier thomas.friedrichsmeier@ruhr-uni-bochum.de, Meik Michalke meik.michalke@uni-duesseldorf.de

See Also

RKTestSuite-class, rktest.makeplugintests

Examples

## Not run: 
result <- rktest.runRKTestSuite()

## End(Not run)

[Package rkwardtests version 0.6.1 Index]