rktest.setSuiteStandards {rkwardtests}	R Documentation

Set RKWard suite standards

Description

Set RKWard plugin test suite standards

Usage

  rktest.setSuiteStandards(suite, testroot=getwd(),
    file=TRUE)

Arguments

suite
Character string naming the test suite to set standards for.

testroot
Path to the test root directory, defaults to the working directory.

file
Logical: If suite is already a present R object, set this to FALSE. Otherwise it is assumed to be a file and fed to source.

Details

Use this function after you plugin passed all tests to set the resulting code, output and R messages as the standard that will be compared to during following tests.

Value

The function simply copies the previously created files from the temporary directory to the directory containing the test standards (inside the testroot).

Author(s)

Thomas Friedrichsmeier thomas.friedrichsmeier@ruhr-uni-bochum.de

See Also

RKTestSuite-class, rktest.makeplugintests

Examples

## Not run: 
rktest.setSuiteStandards("rkward_application_tests.R")

## End(Not run)

[Package rkwardtests version 0.6.1 Index]